fix bug in modal funciton
